requirements.txt
google-genai
pdf2image
pydantic
python-dotenv
pillow
python-magic
